+++ app/app.ts
@@ -1,71 +1,64 @@
-import {App, IonicApp, Nav, Platform} from 'ionic-angular';
-import {StatusBar} from 'ionic-native';
-import { provide } from '@angular/core';
+import { provide, Component, ViewChild } from '@angular/core';
 
-import {PracticePage} from './pages/practice/practice';
-import {QualifyingPage} from './pages/qualifying/qualifying';
-import {RacePage} from './pages/race/race';
-import {ConnectionPage} from './pages/connection/connection';
-import {SettingsPage} from './pages/settings/settings';
-import {AboutPage} from './pages/about/about';
+import { ionicBootstrap, App, Modal, Nav, NavController, Platform } from 'ionic-angular';
 
-import {BLEProvider} from './connections/ble';
-import {DemoProvider} from './connections/demo';
-import {Provider} from './connections/connection';
-import {ControlUnit} from './providers/cu';
-import { Drivers, DefaultDrivers } from './providers/drivers';
-import {Inject, ViewChild} from '@angular/core';
+import { PracticePage, QualifyingPage, RacePage, SettingsPage, ConnectionPage, LoggingPage } from './pages';
+import { ControlUnit, Drivers, DefaultDrivers, Logger, Plugins } from './providers';
+// TODO: move to connections page?
+import { SerialProvider } from './connections/serial';
 
-@App({
-  templateUrl: 'build/app.html',
-  providers: [ControlUnit, provide(Drivers, {useClass: DefaultDrivers})],
-  config: {} // http://ionicframework.com/docs/v2/api/config/Config/
+@Component({
+  providers: [
+    ControlUnit, 
+    provide(Drivers, {useClass: DefaultDrivers}),
+    Logger,
+    Plugins,
+    SerialProvider
+  ],
+  templateUrl: 'build/app.html'
 })
-class MyApp {
+class OpenLapApp {
   rootPage: any = PracticePage;
-  pages: Array<{title: string, component: any}>
+  
+  pages = [
+    { title: 'Practice', component: PracticePage },
+    { title: 'Qualifying', component: QualifyingPage },
+    { title: 'Race', component: RacePage },
+    { title: 'Settings', component: SettingsPage },
+    //{ title: 'Connection', component: ConnectionPage },
+    { title: 'Logging', component: LoggingPage },
+    //{ title: 'About', component: AboutPage }
+  ];
 
   @ViewChild(Nav) nav: Nav;
 
-  constructor(private app: IonicApp, private platform: Platform, private cu: ControlUnit, private drivers:Drivers) {
-          console.log('Drivers: ', this.drivers);
+  constructor(private platform: Platform, private cu: ControlUnit, private drivers: Drivers,
+              private logger: Logger, private plugins: Plugins)
+  {
+      platform.ready().then(() => this.onReady());
 
-    this.initializeApp();
-
-    // used for an example of ngFor and navigation
-    this.pages = [
-      { title: 'Practice', component: PracticePage },
-      { title: 'Qualifying', component: QualifyingPage },
-      { title: 'Race', component: RacePage },
-      //{ title: 'Connection', component: ConnectionPage },
-      //{ title: 'Settings', component: SettingsPage },
-      //{ title: 'About', component: AboutPage }
-    ];
-
-    let provider = platform.is('cordova') ? new BLEProvider() : new DemoProvider();
-
-    provider.scan().subscribe(device => {
-      console.log(device);
-      provider.connect(device).subscribe(connection => {
-        cu.connect(connection);
-        cu.getVersion().subscribe(version => {
-          console.log('CU version', version);
-        });
+      this.plugins.get('insomnia').then(insomnia => {
+          return new Promise((resolve, reject) => {
+              insomnia.keepAwake(resolve, reject);
+          });
+      }).then(() => {
+          this.logger.info('Keeping app awake...');
+      }).catch(error => {
+          this.logger.error('Not using insomnia plugin', error);
       });
-    });
   }
 
-  initializeApp() {
-    this.platform.ready().then(() => {
-      // Okay, so the platform is ready and our plugins are available.
-      // Here you can do any higher level native things you might need.
-      StatusBar.styleDefault();
-      if (window['plugins'] && window['plugins'].insomnia) {
-        window['plugins'].insomnia.keepAwake(() => {
-          console.log('Keeping app awake...');
-        });
-      }
-    });
+  onReady() {
+    // if (window['plugins'] && window['plugins'].insomnia) {
+    //   window['plugins'].insomnia.keepAwake(() => {
+    //     this.logger.info('Keeping app awake...');
+    //   });
+    // }
+    
+    // FIXME: overlay on race screen on first open
+    let modal = Modal.create(ConnectionPage);
+    //modal.onDismiss(settings => {});
+    setTimeout(() => this.nav.present(modal));
   }
 
   openPage(page) {
@@ -73,4 +66,14 @@ class MyApp {
     // we wouldn't want the back button to show in this scenario
     this.nav.setRoot(page.component);
   }
+
+  ionViewLoaded() {
+    console.log('ionViewLoaded');
+    // FIXME: overlay on race screen on first open
+    //let modal = Modal.create(ConnectionPage);
+    //modal.onDismiss(settings => {});
+    //setTimeout(() => this.nav.present(modal));
+  }
 }
+  
+ionicBootstrap(OpenLapApp);