+++ app/components.ts
@@ -1,4 +1,0 @@
-export * from './components/gauge/gauge';
-export * from './components/leaderboard/leaderboard';
-export * from './components/startlight/startlight';
-export * from './components/stripe/stripe';