+++ app/components/fuel-gauge/fuel-gauge.ts
@@ -1,7 +1,7 @@
 import {Component, Input} from '@angular/core';
 
 @Component({
-    selector: 'gauge',
+    selector: 'fuel-gauge',
     template: `
     <span [style.width]="100 * (value - min) / (max - min) + '%'"></span>
     <span></span>
@@ -9,7 +9,7 @@ import {Component, Input} from '@angular/core';
     host: {
     } 
 })
-export class Gauge {
+export class FuelGauge {
     @Input() min: number = 0;
     @Input() max: number = 1.0;
     @Input() value: number;