+++ app/connections/ble.ts
@@ -1,7 +1,11 @@
-import {Connection, Device, Provider} from './connection';
-import {EventEmitter, Injectable} from '@angular/core';
-import {BLE} from 'ionic-native';
+import { Injectable } from '@angular/core';
+
+import { Subject } from 'rxjs/Subject';
 import 'rxjs/add/operator/map';
+
+import { BLE } from 'ionic-native';
+
+import { Connection, Device, ConnectionProvider } from './connection';
 
 const SERVICE_UUID = '39df7777-b1b4-b90b-57f1-7144ae4e4a6a';
 const OUTPUT_UUID = '39df8888-b1b4-b90b-57f1-7144ae4e4a6a';
@@ -9,13 +13,17 @@ const NOTIFY_UUID = '39df9999-b1b4-b90b-57f1-7144ae4e4a6a';
 
 const DOLLAR = '$'.charCodeAt(0);
 
-export class BLEConnection implements Connection {
+export class BLEConnection extends Subject<ArrayBuffer> implements Connection {
 
-  private emitter: any;
+  private subscription: any;
 
+  private logEnabled = false;
+  
   constructor(private peripheral) {
-    console.log('Create BLE connection', peripheral);
-    this.emitter = BLE.startNotification(peripheral.id, SERVICE_UUID, NOTIFY_UUID).map(data => {
+    super();
+    this.log('Creating BLE connection');
+    BLE.startNotification(peripheral.id, SERVICE_UUID, NOTIFY_UUID).map(data => {
+      this.log('Received BLE data', data);
       // strip trailing '$' and prepend missing '0'/'?' for notifications
       let view = new Uint8Array(data);
       if (view[view.length - 1] == DOLLAR) {
@@ -23,37 +31,36 @@ export class BLEConnection implements Connection {
         view[0] = view.length == 6 ? 0x30 : 0x3f;
       }
       return view.buffer;
-    });
-  }
-
-  subscribe(generatorOrNext?: any, error?: any, complete?: any) {
-    this.emitter.subscribe(generatorOrNext, error, complete);
+    }).subscribe(this);
   }
 
   send(data: ArrayBuffer) {
+    this.log('Sending BLE data', data);
     BLE.writeWithoutResponse(this.peripheral.id, SERVICE_UUID, OUTPUT_UUID, data);
   }
 
   close() {
+    this.log('Closing BLE connection');
     BLE.disconnect(this.peripheral.id);
+  }
+  
+  private log(message: string, data?: ArrayBuffer) {
+    if (this.logEnabled) {
+      let now = Date.now();
+      if (data) {
+        console.log(now.toString() + ": " + message + " " + String.fromCharCode.apply(null, new Uint8Array(data)));
+      } else {
+        console.log(now.toString() + ": " + message);
+      }
+    }
   }
 }
 
 @Injectable()
-export class BLEProvider extends Provider {
-  constructor() {
-    super();
-    BLE.enable().then(() => {
-      console.log("Bluetooth is enabled");
-    });
-  }
-
+export class BLEProvider extends ConnectionProvider {
   connect(device: Device) {
-    return BLE.connect(device.id).map(peripheral => new BLEConnection(peripheral));
-  }
-
-  scan() {
-    // TODO: stopScan() when disposed
-    return BLE.startScan([/*SERVICE_UUID*/]);
+    return new Promise((resolve, reject) => {
+      BLE.connect(device.id).subscribe(peripheral => resolve(new BLEConnection(peripheral)), reject);
+    });
   }
 }