+++ app/pages/logging/logging.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+import { Logger } from '../../providers';
+
+@Component({
+  templateUrl: 'build/pages/logging/logging.html',
+})
+export class LoggingPage {
+  constructor(private logger: Logger) {}
+}