+++ CHANGES.rst
@@ -1,3 +1,13 @@
+0.2.0 2016-06-10
+----------------
+
+- Choose connection at startup.
+
+- Add serial connection support.
+
+- Add application logging.
+
+
 0.1.0 2016-06-04
 ----------------